Boat docking and animation logic
Basic boat docking system is now implemented. Animation logic is present, but I believe we need new animations with the same model, or a custom skeletal mesh to get them all working together. Currently, the walking animation is working, and the idle is a "dummy" animation that freezes the mesh. This at least looks intentional :+1: